minor nit :  rename the get bucket location file
