+++ backend/src/api/routes/s3-gateway/commands/get-bucket-location.ts
